third_party/WebKit/LayoutTests/css3/motion-path/property-use-count-offset-rotate.html

<!DOCTYPE html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>

<div id="target" class="p"></div>
<script>
test(function() {
  assert_false(internals.isCSSPropertyUseCounted(document, "offset-rotate"));

  var styleElement = document.createElement('style');
  styleElement.textContent = ".p { offset-rotate: 0deg; }";
  document.head.appendChild(styleElement);
  getComputedStyle(target);

  assert_true(internals.isCSSPropertyUseCounted(document, "offset-rotate"));
});
</script>
